requirements.txt
Django==3.0
djangosaml2idp==0.7.2
